requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-environ==0.10.0
django-storages==1.13.2
djangorestframework==3.14.0
drf-spectacular==0.26.2
frozenlist==1.3.3
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
multidict==6.0.4
phonenumbers==8.13.13
PyJWT==2.7.0
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
requests==2.31.0
sqlparse==0.4.4
twilio==8.2.2
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==2.0.3
yarl==1.9.2
